pyproject.toml
[tool.poetry]
name = "image-tools-api"
version = "1.0.0"
description = "A REST API server for image processing."
authors = ["James Charles <aquarokk@gmail.com>"]
license = "MIT"
readme = "README.md"
keywords = ["pillow", "rest-api", "image-processing"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.66.1"
uvicorn = {extras = ["standard"], version = "^0.14.0"}
Pillow = "^8.3.1"
python-multipart = "^0.0.5"
colorific = "^0.3"
scikit-image = "^0.18.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
black = "^21.7b0"
pytest-cov = "^2.12.1"
requests = "^2.26.0"
opencv-python = "^4.5.3"
aiohttp = "^3.7.4"
aiofiles = "^0.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-vv --tb=short -p no:warnings"
testpaths = "tests"
xfail_strict = true